@andrewballantyne code review comments for SDKK8sReducers
