+++ packages/lib-utils/src/types/redux.ts
@@ -1,4 +1,6 @@
 import type { Map as ImmutableMap } from 'immutable';
+import type { AnyAction } from 'redux';
+import type { ThunkDispatch } from 'redux-thunk';
 import type { K8sResourceCommon } from './k8s';
 
 export type K8sState = ImmutableMap<string, unknown> | undefined;
@@ -14,11 +16,13 @@ export type ImpersonateKind = {
   subprotocols: string[];
 };
 
-export type CoreState = {
-  activeCluster?: string;
-  user?: UserKind;
-  impersonate?: ImpersonateKind;
-};
+export type CoreState =
+  | {
+      activeCluster?: string;
+      user?: UserKind;
+      impersonate?: ImpersonateKind;
+    }
+  | undefined;
 
 export type SDKStoreState = {
   sdkCore: CoreState;
@@ -28,3 +32,9 @@ export type SDKStoreState = {
 export type GetImpersonate = (state: SDKStoreState) => ImpersonateKind | undefined;
 
 export type GetCluster = (state: SDKStoreState) => string;
+
+export type DispatchWithThunk = ThunkDispatch<SDKStoreState, undefined, AnyAction>;
+
+export type GetState = () => SDKStoreState;
+
+export type GetUser = (state: SDKStoreState) => UserKind | undefined;
